Addressed review comments.Moved main folder out of airport package.
